add generate controller command

+++ tests/generate.test.js
@@ -6,17 +6,28 @@ const cliPath = path.join(__dirname, '../app/cli.js');
 const testDirPath = path.join(__dirname, '../test-folder');
 
 describe('`generate` command', () => {
-    beforeAll(() => {
-        fs.rmSync(testDirPath, { recursive: true, force: true });
+    beforeEach(() => {
+        if (fs.existsSync(testDirPath)) {
+            fs.rmSync(testDirPath, { recursive: true, force: true });
+        }
         fs.mkdirSync(testDirPath);
     });
-    afterAll(() => {
-        fs.rmSync(testDirPath, { recursive: true, force: true });
+    afterEach(() => {
+        if (fs.existsSync(testDirPath)) {
+            fs.rmSync(testDirPath, { recursive: true, force: true });
+        }
     });
 
     test('`generate dto` command should generate all DTO files', (done) => {
         execFile('node', [cliPath, 'generate', 'dto', 'test1', './test-folder'], () => {
             expect(fs.readdirSync(testDirPath).length).toEqual(9)
+            done()
+        });
+    });
+
+    test('`generate controller` command should generate controller', (done) => {
+        execFile('node', [cliPath, 'generate', 'controller', 'test1', './test-folder'], () => {
+            expect(fs.readdirSync(testDirPath).length).toEqual(1);
             done()
         });
     });
